cleanup: colocate shared type exports in src/types/
Moves RadioProgressPhase, RadioProgress, and RadioState types from hook files to src/types/radio.ts for better organization and single-source-of-truth.

- Added RadioState to src/types/radio.ts (previously in src/hooks/useRadio.ts)
- Updated all component imports to use src/types/radio instead of hook files
- Updated hook re-exports to import from src/types/radio
- RadioProgressPhase and RadioProgress already existed in radio.ts but were being re-exported from hooks
- All imports now properly reference the type definitions in src/types/

Addresses issue #425
