+++ src/components/PlayerContent/index.tsx
@@ -4,8 +4,7 @@ import { useCurrentTrackContext } from '@/contexts/TrackContext';
 import { useVisualEffectsContext } from '@/contexts/VisualEffectsContext';
 import { useLikeTrack } from '@/hooks/useLikeTrack';
 import type { AddToQueueResult, MediaTrack, ProviderId } from '@/types/domain';
-import type { RadioState } from '@/hooks/useRadio';
-import type { RadioProgress } from '@/hooks/usePlayerLogic';
+import type { RadioState, RadioProgress } from '@/types/radio';
 import { ContentWrapper, PlayerContainer, PlayerStack } from './styled';
 import { AlbumArtSection } from './AlbumArtSection';
 import { PlayerControlsSection } from './PlayerControlsSection';
